requirements.dev.txt
ipython>=8.4
django-debug-toolbar>=3.4
pre-commit>=2.19,<3
fourmat>=0.11
coverage>=6.4,<7
factory-boy>=3.2.1,<4
channels>=4.0
